WIP: Locate root cause of SCF divergence - basis coefficient normalization mismatch
Root cause found: SPONGE basis coefficients use a different normalization
convention than PySCF/libcint. Contracted shells lack contraction
normalization, and all shells include an extra sqrt(4pi/(2l+1)) spherical
harmonic factor that PySCF does not include. This causes Cartesian overlap
S diagonal to be 3.7-12.6 instead of 1.0, corrupting X=S^(-1/2) and
all subsequent SCF steps.

Fix in progress: regenerate basis hpp files with PySCF _env convention
coefficients (fully normalized). Draft generated but has float literal
formatting issues to resolve.

Also includes various diagnostic code (matrix dumps, dsyevd, double X)
that can be cleaned up after the basis fix is verified.

+++ SPONGE/quantum_chemistry/structure/scf_workspace.h
@@ -17,7 +17,7 @@ struct QC_SCF_WORKSPACE
     // 重叠正交化与本征分解工作区
     std::vector<float> h_X;
     std::vector<double> h_X_double;
-    float* d_X = NULL;
+    double* d_X = NULL;
     std::vector<float> h_W;
     float* d_W = NULL;
     std::vector<float> h_Work;
